New structure, annotations changing, adding yml file
