+++ pom.xml
@@ -8,7 +8,6 @@
     <artifactId>moscycling</artifactId>
     <version>1.0-SNAPSHOT</version>
 
-
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
@@ -27,86 +26,13 @@
         <dependency>
             <groupId>org.postgresql</groupId>
             <artifactId>postgresql</artifactId>
-            <version>42.2.5</version>
         </dependency>
-
-<!--Лишняя зависимость. Шаблонизатор не нужен. Папка templates пустая.  -->
-        <!--<dependency>-->
-        <!--<groupId>org.springframework.boot</groupId>-->
-        <!--<artifactId>spring-boot-starter-thymeleaf</artifactId>-->
-        <!--</dependency>-->
-
-<!--Лишняя зависимость. В этом проекте возможности devtools не задействованы-->
-        <!--<dependency>-->
-        <!--<groupId>org.springframework.boot</groupId>-->
-        <!--<artifactId>spring-boot-devtools</artifactId>-->
-        <!--<optional>true</optional>-->
-        <!--</dependency>-->
-
-
-<!--Дублирующая зависимость. Spring Boot parent POM содержит в себе зависимость jackson-databind-->
-        <!-- https://mvnrepository.com/artifact/com.fasterxml.jackson.core/jackson-core -->
-        <!--<dependency>-->
-        <!--<groupId>com.fasterxml.jackson.core</groupId>-->
-        <!--<artifactId>jackson-databind</artifactId>-->
-        <!--<version>2.9.9</version>-->
-        <!--</dependency>-->
-
-
-        <!-- https://mvnrepository.com/artifact/postgresql/postgresql -->
-
-
-<!--Дублирующая зависимость.Hikari is the default DataSource implementation. spring-boot-starter-jdbc POM
-содержит эту зависимость. А spring-boot-starter-data-jpa сожержит spring-boot-starter-jdbc-->
-        <!--&lt;!&ndash; https://mvnrepository.com/artifact/com.zaxxer/HikariCP &ndash;&gt;-->
-        <!--<dependency>-->
-        <!--<groupId>com.zaxxer</groupId>-->
-        <!--<artifactId>HikariCP</artifactId>-->
-        <!--<version>3.3.1</version>-->
-        <!--</dependency>-->
-
-<!--Зависимость лишняя. До тестирования проекта руки так и не дошли. Пока Spring приложения тестировать не умею.
-Зависимость была включена безосновательно-->
-        <!--<dependency>-->
-        <!--<groupId>junit</groupId>-->
-        <!--<artifactId>junit</artifactId>-->
-        <!--<version>4.12</version>-->
-        <!--<scope>test</scope>-->
-        <!--<exclusions>-->
-        <!--<exclusion>-->
-        <!--<groupId>org.hamcrest</groupId>-->
-        <!--<artifactId>hamcrest-core</artifactId>-->
-        <!--</exclusion>-->
-        <!--</exclusions>-->
-        <!--</dependency>-->
-
- <!--Зависимость лишняя. Тесты к проекту не написаны-->
-        <!--<dependency>-->
-        <!--<groupId>org.hamcrest</groupId>-->
-        <!--<artifactId>hamcrest-library</artifactId>-->
-        <!--<version>1.3</version>-->
-        <!--<scope>test</scope>-->
-        <!--</dependency>-->
-
- <!--Зависимость лишняя. В проекте thymeleaf не используется-->
-        <!--<dependency>-->
-        <!--<groupId>org.thymeleaf</groupId>-->
-        <!--<artifactId>thymeleaf</artifactId>-->
-        <!--<version>3.0.9.RELEASE</version>-->
-        <!--</dependency>-->
-        <!--<dependency>-->
-        <!--<groupId>org.thymeleaf</groupId>-->
-        <!--<artifactId>thymeleaf-spring4</artifactId>-->
-        <!--<version>3.0.9.RELEASE</version>-->
-        <!--</dependency>-->
-
     </dependencies>
 
     <properties>
         <java.version>1.8</java.version>
 
     </properties>
-
 
     <build>
         <plugins>
