+++ src/main/java/ru/arkaleks/moscycling/Application.java
@@ -1,30 +1,13 @@
 package ru.arkaleks.moscycling;
 
-
-import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-
-import java.util.*;
 
 
-@SpringBootApplication(scanBasePackages = {"ru.arkaleks.moscycling"})
+@SpringBootApplication
 public class Application {
 
     public static void main(String[] args) {
         SpringApplication.run(Application.class, args);
     }
-
-
-    @Bean
-    CommandLineRunner initDatabase(CyclePathRepository repository) {
-        return args -> {
-            List<CyclePath> temp = CyclePathConvert.getInstance().getCyclePathDataFromOpenSource();
-            for (CyclePath cp : temp) {
-                repository.save(cp);
-            }
-
-        };
     }
-}
