--- src/main/java/ru/arkaleks/moscycling/CustomGlobalExceptionHandler.java
@@ -1,22 +1,0 @@
-//package ru.arkaleks.moscycling;
-//
-//import org.springframework.http.HttpStatus;
-//import org.springframework.web.bind.annotation.ControllerAdvice;
-//import org.springframework.web.bind.annotation.ExceptionHandler;
-//import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-//
-//import javax.servlet.http.HttpServletResponse;
-//import java.io.IOException;
-//
-//@ControllerAdvice
-//public class CustomGlobalExceptionHandler extends ResponseEntityExceptionHandler {
-//    @ExceptionHandler(CyclePathNotFoundException.class)
-//    public void springHandleNotFound(HttpServletResponse response) throws IOException {
-//        response.sendError(HttpStatus.NOT_FOUND.value());
-//    }
-//
-//    @ExceptionHandler(CyclePathUnSupportedFieldPatchException.class)
-//    public void springUnSupportedFieldPatch(HttpServletResponse response) throws IOException {
-//        response.sendError(HttpStatus.METHOD_NOT_ALLOWED.value());
-//    }
-//}
