+++ src/main/java/ru/arkaleks/moscycling/model/GeoData.java
@@ -1,4 +1,4 @@
-package ru.arkaleks.moscycling;
+package ru.arkaleks.moscycling.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
