+++ src/main/java/ru/arkaleks/moscycling/repository/CellsRepository.java
@@ -1,6 +1,7 @@
-package ru.arkaleks.moscycling;
+package ru.arkaleks.moscycling.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import ru.arkaleks.moscycling.model.Cells;
 
 public interface CellsRepository extends JpaRepository<Cells, Integer> {
 }
