+++ src/main/java/ru/arkaleks/moscycling/repository/CyclePathRepository.java
@@ -1,6 +1,7 @@
-package ru.arkaleks.moscycling;
+package ru.arkaleks.moscycling.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import ru.arkaleks.moscycling.model.CyclePath;
 
 public interface CyclePathRepository extends JpaRepository<CyclePath, Integer> {
 }
