+++ src/main/java/ru/arkaleks/moscycling/repository/GeoDataRepository.java
@@ -1,6 +1,7 @@
-package ru.arkaleks.moscycling;
+package ru.arkaleks.moscycling.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import ru.arkaleks.moscycling.model.GeoData;
 
 public interface GeoDataRepository extends JpaRepository<GeoData, Integer> {
 }
