--- src/main/resources/application.properties
@@ -1,25 +1,0 @@
-spring.datasource.hikari.connectionTimeout=20000
-spring.datasource.hikari.maximumPoolSize=5
-
-//logging.pattern.console=%-5level %logger{36} - %msg%n
-
-## PostgreSQL
-#spring.datasource.url=jdbc:postgresql://192.168.1.4:5432/postgres
-
-
-server.port=9090
-spring.jpa.database=POSTGRESQL
-spring.datasource.platform=postgres
-spring.datasource.url=jdbc:postgresql://localhost:5432/moscycling
-spring.datasource.username=postgres
-spring.datasource.password=p091ulp
-
-#drop n create table again, good for testing, comment this in production
-spring.jpa.hibernate.ddl-auto=create-drop
-
-spring.jpa.show-sql=true
-
-spring.jpa.generate-ddl=true
-
-
-#spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation=true
