+++ src/main/resources/application.yml
@@ -1,0 +1,17 @@
+spring:
+  jpa:
+    generate-ddl: true
+    show-sql: true
+    hibernate:
+      ddl-auto: create-drop
+  datasource:
+    hikari:
+      connectionTimeout: 20000
+      maximumPoolSize: 5
+    platform: postgres
+    url: jdbc:postgresql://${DB_HOST:localhost}:${DB_PORT:5432}/${DB_NAME:accounts}
+    username: ${DB_LOGIN:postgres}
+    password: ${DB_PASSWORD:p091ulp}
+    driver-class-name: org.postgresql.Driver
+server:
+  port: 9090
